+++ lib/ap_compat/AP_HAL/AP_HAL_Namespace.h
@@ -1,0 +1,12 @@
+/**
+ * @file AP_HAL_Namespace.h
+ * @brief AP_HAL namespace forward declarations
+ */
+
+#pragma once
+
+namespace AP_HAL {
+
+class HAL;
+
+} // namespace AP_HAL