+++ lib/ap_compat/AP_HAL/Semaphores.h
@@ -1,0 +1,13 @@
+/**
+ * @file Semaphores.h
+ * @brief HAL_Semaphore stub for RocketChip
+ */
+
+#pragma once
+
+#include <cstdint>
+
+// HAL_Semaphore is defined in AP_HAL_Compat.h
+// This header just re-exports it for code that includes AP_HAL/Semaphores.h
+
+#include "../AP_HAL_Compat.h"