+++ lib/ap_compat/AP_Logger/AP_Logger_config.h
@@ -1,0 +1,9 @@
+/**
+ * @file AP_Logger_config.h
+ * @brief AP_Logger stub for RocketChip
+ */
+
+#pragma once
+
+#define HAL_LOGGING_ENABLED 0
+#define AP_LOGGER_ENABLED 0